Fixes auto-retry-delay type
`auto-retry-delay` was set to be u16, but its default value was beyond u16::MAX.

+++ watchtower-plugin/src/main.rs
@@ -625,7 +625,7 @@ async fn main() -> Result<(), Error> {
         e
     })?;
 
-    let auto_retry_delay = u16::try_from(
+    let auto_retry_delay = u32::try_from(
         midstate
             .option(constants::WT_AUTO_RETRY_DELAY)
             .unwrap()
